+++ .gitignore
@@ -14,8 +14,6 @@ dist-ssr
 coverage
 *.local
 
-.env
-
 /cypress/videos/
 /cypress/screenshots/
