+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" href="src/assets/img/suitcases.png" />
+    <link rel="icon" href="/suitcases.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>TripEasy 旅遊電商</title>
