+++ package.json
@@ -11,13 +11,13 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "bootstrap": "^5.3.5",
     "pinia": "^3.0.1",
-    "sweetalert2": "^11.6.13",
+    "sweetalert2": "^11.21.0",
     "vue": "^3.5.13",
     "vue-axios": "^3.5.2",
-    "vue-router": "^4.5.1"
+    "vue-router": "^4.5.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.22.0",
