+++ src/views/HomeView.vue
@@ -1,0 +1,6 @@
+<script setup>
+</script>
+
+<template>
+
+</template>
