+++ src/views/admin/collectPage.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="about">
+    <h1>這是個人收藏頁面</h1>
+  </div>
+</template>
+
+<style>
+@media (min-width: 1024px) {
+  .about {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+  }
+}
+</style>
