+++ src/views/admin/manage/websiteSetting.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="about">
+    <h1>這是網站管理頁面</h1>
+  </div>
+</template>
+
+<style>
+@media (min-width: 1024px) {
+  .about {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+  }
+}
+</style>
