+++ src/views/admin/memberPage.vue
@@ -1,0 +1,7 @@
+<template>
+  <router-view /> <!-- 🔥 這一行是必要的，顯示子路由 -->
+</template>
+
+<style>
+
+</style>
